move cross_sphere_line_test method into test_controlmath.cpp remove mc_pos_control_tests since empty move test_controlmath.cpp into systemcmds/tests directory
